+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * `Common`:
     * `applyDiff` will return `Diff` object since this release
+    * `Android`:
+      * New functions/extensions `findViewsByTag` and `findViewsByTagInActivity`
 * `Coroutines`:
     * Add `Flow` extensions `flatMap`, `flatMapNotNull` and `flatten`
     * Add `Flow` extensions `takeNotNull` and `filterNotNull`